atoi, string operations

+++ src/ft_strlen.c
@@ -6,7 +6,7 @@
 /*   By: ydidenko <ydidenko@student.codam.nl>         +#+                     */
 /*                                                   +#+                      */
 /*   Created: 2023/10/06 13:32:13 by ydidenko      #+#    #+#                 */
-/*   Updated: 2023/10/17 19:34:50 by ydidenko      ########   odam.nl         */
+/*   Updated: 2023/10/23 19:07:48 by ydidenko      ########   odam.nl         */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -19,7 +19,7 @@
 
 size_t	ft_strlen(char *str)
 {
-	int	i;
+	size_t	i;
 
 	i = 0;
 	while (str[i])
